fix: internal выделить middleware  в отдельный пакет от handlers
